+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* fix table title text justification
 * `--build-dir` argument
 * new `--use-custom-html` arg, `--all-other-files` site now includes biographical panels
 * If --all-other-files is enables, print the biographical panels, otherwise just print a big table